Rename search_course directory to search_subject_v0
