+++ lib/feature/search_subject_v0/domain/search_course_filter_options.dart
@@ -1,4 +1,4 @@
-import 'package:dotto/feature/search_course/domain/search_course_filter_option_choice.dart';
+import 'package:dotto/feature/search_subject_v0/domain/search_course_filter_option_choice.dart';
 
 enum SearchCourseFilterOptions {
   // 学部・大学院
