+++ lib/feature/search_subject_v0/repository/course_filter_extractor.dart
@@ -1,5 +1,5 @@
-import 'package:dotto/feature/search_course/domain/search_course_filter_option_choice.dart';
-import 'package:dotto/feature/search_course/domain/search_course_filter_options.dart';
+import 'package:dotto/feature/search_subject_v0/domain/search_course_filter_option_choice.dart';
+import 'package:dotto/feature/search_subject_v0/domain/search_course_filter_options.dart';
 
 final class CourseFilterExtractor {
   static CourseFilterData extractFilters(
