+++ lib/feature/search_subject_v0/repository/syllabus_database_config.dart
