+++ lib/feature/search_subject_v0/widget/search_course_checkbox_item.dart
