+++ lib/feature/search_subject_v0/widget/search_course_filter_checkbox.dart
@@ -1,6 +1,6 @@
-import 'package:dotto/feature/search_course/domain/search_course_filter_option_choice.dart';
-import 'package:dotto/feature/search_course/domain/search_course_filter_options.dart';
-import 'package:dotto/feature/search_course/widget/search_course_checkbox_item.dart';
+import 'package:dotto/feature/search_subject_v0/domain/search_course_filter_option_choice.dart';
+import 'package:dotto/feature/search_subject_v0/domain/search_course_filter_options.dart';
+import 'package:dotto/feature/search_subject_v0/widget/search_course_checkbox_item.dart';
 import 'package:flutter/material.dart';
 
 final class SearchCourseFilterCheckbox extends StatelessWidget {
