+++ lib/feature/search_subject_v0/widget/search_course_result_section.dart
@@ -1,5 +1,5 @@
-import 'package:dotto/feature/search_course/search_course_viewmodel.dart';
-import 'package:dotto/feature/search_course/widget/search_course_result.dart';
+import 'package:dotto/feature/search_subject_v0/search_course_viewmodel.dart';
+import 'package:dotto/feature/search_subject_v0/widget/search_course_result.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
